+++ src/pools/selection-strategies/less-busy-worker-choice-strategy.ts
@@ -1,18 +1,24 @@
 import type { IPoolWorker } from '../pool-worker'
 import { AbstractWorkerChoiceStrategy } from './abstract-worker-choice-strategy'
+import type { RequiredStatistics } from './selection-strategies-types'
 
 /**
- * Selects the less used worker.
+ * Selects the less busy worker.
  *
  * @typeParam Worker - Type of worker which manages the strategy.
  * @typeParam Data - Type of data sent to the worker. This can only be serializable data.
  * @typeParam Response - Type of response of execution. This can only be serializable data.
  */
-export class LessUsedWorkerChoiceStrategy<
+export class LessBusyWorkerChoiceStrategy<
   Worker extends IPoolWorker,
   Data,
   Response
 > extends AbstractWorkerChoiceStrategy<Worker, Data, Response> {
+  /** {@inheritDoc} */
+  public readonly requiredStatistics: RequiredStatistics = {
+    runTime: true
+  }
+
   /** {@inheritDoc} */
   public reset (): boolean {
     return true
@@ -20,20 +26,19 @@ export class LessUsedWorkerChoiceStrategy<
 
   /** {@inheritDoc} */
   public choose (): Worker {
-    let minNumberOfTasks = Infinity
-    let lessUsedWorker!: Worker
+    let minRunTime = Infinity
+    let lessBusyWorker!: Worker
     for (const value of this.pool.workers.values()) {
       const worker = value.worker
-      const tasksUsage = this.pool.getWorkerTasksUsage(worker)
-      const workerTasks =
-        (tasksUsage?.run as number) + (tasksUsage?.running as number)
-      if (!this.isDynamicPool && workerTasks === 0) {
+      const workerRunTime = this.pool.getWorkerTasksUsage(worker)
+        ?.runTime as number
+      if (!this.isDynamicPool && workerRunTime === 0) {
         return worker
-      } else if (workerTasks < minNumberOfTasks) {
-        minNumberOfTasks = workerTasks
-        lessUsedWorker = worker
+      } else if (workerRunTime < minRunTime) {
+        minRunTime = workerRunTime
+        lessBusyWorker = worker
       }
     }
-    return lessUsedWorker
+    return lessBusyWorker
   }
 }