+++ CHANGELOG.md
@@ -7,11 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add `LESS_BUSY` worker choice strategy.
+
 ### Changed
 
 - Optimize worker storage in pool.
 - Optimize worker alive status check.
-- Rename worker choice strategy `LESS_RECENTLY_USED to `LESS_USED` .
+- Rename worker choice strategy `LESS_RECENTLY_USED` to `LESS_USED`.
 - Optimize `LESS_USED` worker choice strategy.
 
 ### Fixed